+++ src/content/js/content_script.js
@@ -502,9 +502,9 @@
                 terms: ['rottentomatoes.com/m']
             },
             icon: {
-                containerSelector: 'h1',
+                containerSelector: '#hero-wrap [context="heading"]',
                 locator: 'prepend',
-                imgStyles: 'width: 35px; margin: -8px 10px 0 0;'
+                imgStyles: 'width: 30px; margin: -4px 10px 0 0;'
             }
         },
         {
@@ -525,9 +525,9 @@
                 terms: ['rottentomatoes.com/tv']
             },
             icon: {
-                containerSelector: 'h1',
+                containerSelector: '#hero-wrap [context="heading"]',
                 locator: 'prepend',
-                imgStyles: 'width: 35px; margin: -8px 10px 0 0;'
+                imgStyles: 'width: 30px; margin: -4px 10px 0 0;'
             }
         },
         {
@@ -554,7 +554,7 @@
                 modifiers: [
                     {
                         type: 'replace',
-                        from: ' Reviews',
+                        from: ' reviews',
                         to: ''
                     }
                 ]