+++ tests/playwright/site.integrations.tests/tvdb.spec.ts
@@ -6,12 +6,12 @@ import { getExpectedRadarrUrl, getExpectedSonarrUrl } from '../helpers';
 test('tvdb tv has sonarr icon', async ({ page }) => {
   await page.goto('https://www.thetvdb.com/series/fringe', { waitUntil: 'commit' });
   await expect(page.locator(iconDataLocator)).toHaveCount(1);
-  await expect(page.locator(iconDataLocator)).toHaveAttribute('href', getExpectedSonarrUrl('tvdb:82066'));
+  await expect(page.locator(iconDataLocator)).toHaveAttribute('href', getExpectedSonarrUrl('tvdb:82066'), { ignoreCase: true });
 });
 
 // tvdb id doesn't work with radarr
 test('tvdb movie has radarr icon', async ({ page }) => {
   await page.goto('https://thetvdb.com/movies/the-dark-knight', { waitUntil: 'commit' });
   await expect(page.locator(iconDataLocator)).toHaveCount(1);
-  await expect(page.locator(iconDataLocator)).toHaveAttribute('href', getExpectedRadarrUrl('The%20Dark%20Knight'));
+  await expect(page.locator(iconDataLocator)).toHaveAttribute('href', getExpectedRadarrUrl('The%20Dark%20Knight'), { ignoreCase: true });
 });